Unify workers/threads type
blkhash_new() accept unsigned, but config_init() and struct config used
an int. Since this is a size value, unsigned value makes more sense.

There is one ugly issue - when initializing the last_index using

    id - config.workers

We don't get the expected value, likely because id is promoted to
unsigned int. Fixed by casting config.workers to int.

Part-of: #79

+++ lib/worker.c
@@ -166,7 +166,7 @@ int worker_init(struct worker *w, int id, const struct config *config)
 
     w->id = id;
     w->config = config;
-    w->last_index = id - config->workers;
+    w->last_index = id - (int)config->workers;
     w->queue_len = 0;
     w->error = 0;
     w->running = true;
